fix: build after heavy datamodel refactor

+++ Microservices/ListingsMicroservice/Services/MarketTrends/IMarketTrendService.cs
@@ -4,7 +4,7 @@
     {
         decimal GetAveragePrice();
 
-        double GetAveragePricePerSquareMeter();
+        decimal GetAveragePricePerSquareMeter();
 
 
     }
